% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFunctions.R
\name{assessSignatures}
\alias{assessSignatures}
\title{Main function for fitting signatures by group with confidence intervals.}
\usage{
assessSignatures(DNMs, cancer_signatures = cancer_signatures,
  group = "parent", rnd.seed = 1234, n = 1000)
}
\arguments{
\item{DNMs}{A data.frame that contains information about the mutations.
Necessary columns are \code{surrounding}, \code{substitution} plus a column  with information about the groups.}

\item{cancer_signatures}{A matrix with relative abundances of all 96 mutation types in the rows and all signatures to be investigated in the columns.}

\item{group}{Name of the column in \code{DNMs} to base the groups upon.}

\item{rnd.seed}{Random seed to involve in the resampling of the fitting.}

\item{n}{Number of resamplings to be done.}
}
\description{
Main function for fitting signatures by group with confidence intervals.
}
